Item Keluar
Transaksi Item Keluar
TesCase (CrudTest dan HttpTest)

+++ app/Observers/KategoriTransaksiObserver.php
@@ -6,13 +6,14 @@ use App\KategoriTransaksi;
 use App\KasMasuk;
 use App\KasKeluar;
 use Session;
+use Auth;
 
 class KategoriTransaksiObserver
 {
     public function deleting(KategoriTransaksi $KategoriTransaksi)
     {       
-        $data_kategori_masuk = KasMasuk::where('kategori', $KategoriTransaksi->id)->count();
-        $data_kategori_keluar = KasKeluar::where('kategori', $KategoriTransaksi->id)->count();
+        $data_kategori_masuk = KasMasuk::where('kategori', $KategoriTransaksi->id)->where('id_warung', Auth::user()->id_warung)->count();
+        $data_kategori_keluar = KasKeluar::where('kategori', $KategoriTransaksi->id)->where('id_warung', Auth::user()->id_warung)->count();
 
     	if ($data_kategori_masuk > 0 OR $data_kategori_keluar > 0) {
     		$pesan_alert = 
@@ -39,7 +40,7 @@ class KategoriTransaksiObserver
                     <b>Sukses : Kategori Transaksi Berhasil Dihapus</b>
             </div>';
 
-            KategoriTransaksi::where('id', $KategoriTransaksi->id)->delete();
+            KategoriTransaksi::where('id', $KategoriTransaksi->id)->where('id_warung', Auth::user()->id_warung)->delete();
 
             Session:: flash("flash_notification", [
                     "level"=>"success",
